My code review suggestions.
