+++ app/src/main/java/com/locoquest/app/RetrofitClientInstance.kt
@@ -2,10 +2,11 @@ package com.locoquest.app
 
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
+import retrofit2.converter.moshi.MoshiConverterFactory
 
-class RetrofitClientInstance {
+object RetrofitClientInstance {
     private var retrofit: Retrofit? = null;
-    private val BASE_URL = ""//TODO
+    private val BASE_URL = "https://geodesy.noaa.gov/api/nde/"
 
     val retrofitInstance : Retrofit?
         get() {
